update timeout in CI
