+++ .github/workflows/testing-go.yml
@@ -33,9 +33,6 @@ jobs:
           - 'transformers'
           - 'netutils'
           - 'processors'
-        # exclude:
-        #   - os-version: macos-latest
-        #     go-version: '1.20'
     
     runs-on: ${{ matrix.os-version }}
 
@@ -63,7 +60,7 @@ jobs:
         sudo go version
         
     - name: Test ${{ matrix.package }}
-      run: sudo go test -timeout 120s ./${{ matrix.package }}/ -race -cover -v
+      run: sudo go test -timeout 240s ./${{ matrix.package }}/ -race -cover -v
       
   int:
     runs-on: ubuntu-22.04
